Fixed Bug in custom-header.php
In the check that determines whether  the Custom Header is output on an Archive template, there was a call to is_shop being used without a check for WooCommerce class. The condition now triggers if WooCommerce is active and it IS NOT the Shop, or if WooCommerce doesnt exist at all.